requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
autocorrect==2.6.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.6
gingerit==0.0.0.1
greenlet==3.1.1
groq==0.14.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
language_tool_python==2.8.2
nltk==3.9.1
pydantic==2.10.5
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
